The whole project
The project consists of 3 controllers for each tier respectively. Tier 1
receives the data from the user and sends it to Tier2, via RMI. Tier2
receives the data and processes it into 'Tray's object and sends it to
tier3. Tier3 packages the data and uploads it to the Database.
Needs : JDBC controller